61st Commit
Unbeknown to me, for I suspect a long time now, admob interstitial ads are required to include *rich media* which is a short version of video ads. I do not want video ads no matter how short.

Really painful. I only wanted to show ads that are instantly dismissible.

Ads are entirely disabled as of the previous commit, but lowering the threshold and probability of interstitial ads being displayed for the future.